notification: Fix timeouts for error messages
When a message is sent by a plugin, it has a log level, which the
notification plugin uses to indicate what urgency level to set.
Messages with Fatal and Error log levels were being set as critical
urgency, which means the timeout was not respected. Here we make it
so critical is only set for Fatal, and allows timeouts to work for
Error messages.