grid: Restructure edge handling to trigger action on edges between outputs
Output selection was buggy and addressed in compiz core. Without the core
changes to outputDeviceForGeometry(), windows that were grabbed but the
center was offscreen, it would default to output 0 which might not be the
expected one.
The code selects the output that the grabbed window is predominately overlapping.
By contrast, it does not use the mouse pointer location to select the target
output, which might make usage more intuitive. Closes #19.