Fix combination of showmouse and ezoom
The ordering of showmouse and ezoom needs to be deterministic to get
coherent behavior.

Then, ezoom mangles transform according to the zoom. showmouse must thus
follow it, otherwise it will show the mouse position as it is in the X frame
buffer instead of as it is on the actual rendering.